Modified index.html
The logo was spreading beyond the size of the page due to its unlimited size. This creates a bottom scroll bar to allow full accessibility to the entire content on the page. 
By limiting the logo's maximum width to the width of the parent element, the horizontal scroll bar is dealt with and the user experience is much better.